requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.44.0
numpy
matplotlib
tqdm
pyyaml
wandb
